Move edit form to own component
