+++ apps/backend/src/api/projects.route.js
@@ -48,7 +48,7 @@ const updateProject = (req, res) => {
     { new: true },
     (err, Project) => {
       if (err) {
-        res.sed(err);
+        res.send(err);
       } else res.json(Project);
     }
   );
